Fixed migration file

+++ database/migrations/2019_06_22_164123_game_extra_fields.php
@@ -13,14 +13,24 @@ class GameExtraFields extends Migration
      */
     public function up()
     {
-        Schema::table('games', function (Blueprint $table) {
-            $table->string('start_date')->nullable();
-            $table->integer('current_server')->default(0);
+	    if ( ! Schema::hasColumn( 'games', 'start_date' ) ) {
+		    Schema::table('games', function (Blueprint $table) {
+			    $table->string('start_date')->nullable();
+		    });
+	    }
 
-            $table->dropColumn('game_duration');
+	    if ( Schema::hasColumn( 'games', 'game_duration' ) ) {
+		    Schema::table('games', function (Blueprint $table) {
+			    $table->dropColumn('game_duration');
+		    });
+	    }
 
-            $table->foreign('current_server')->references('id')->on('users');
-        });
+	    if ( ! Schema::hasColumn( 'games', 'current_server' ) ) {
+		    Schema::table('games', function (Blueprint $table) {
+			    $table->integer('current_server')->default(0);
+			    $table->foreign('current_server')->references('id')->on('users');
+		    });
+	    }
     }
 
     /**
@@ -30,16 +40,23 @@ class GameExtraFields extends Migration
      */
     public function down()
     {
-        Schema::table('games', function (Blueprint $table) {
-            $table->integer('game_duration')->nullable();
-        });
+    	if ( ! Schema::hasColumn( 'games', 'game_duration' ) ) {
+		    Schema::table('games', function (Blueprint $table) {
+			    $table->integer('game_duration')->nullable();
+		    });
+	    }
 
-        Schema::table('games', function (Blueprint $table) {
-            $table->dropColumn('start_date');
-        });
 
-        Schema::table('games', function (Blueprint $table) {
-            $table->dropColumn('current_server');
-        });
+	    if ( Schema::hasColumn( 'games', 'start_date' ) ) {
+		    Schema::table('games', function (Blueprint $table) {
+			    $table->dropColumn('start_date');
+		    });
+	    }
+
+	    if ( Schema::hasColumn( 'games', 'current_server' ) ) {
+	        Schema::table('games', function (Blueprint $table) {
+	            $table->dropColumn('current_server');
+	        });
+	    }
     }
 }
